finala2e/requirements.txt
langchain-google-genai>=0.0.3
langchain-core>=0.1.0
langgraph>=0.0.20
httpx>=0.24.0
pydantic>=2.0.0
starlette>=0.30.0
sse-starlette>=1.6.0
uvicorn>=0.23.0
jwcrypto>=1.5.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
click>=8.1.0
mcp>=0.1.0
langchain-mcp-adapters>=0.0.1 